Crypto_LLM/requirements.txt
ccxt>=2.0.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
requests>=2.28.0
python-dotenv>=0.21.0
openai
ollama
uv
litellm[proxy]
chromadb>=0.4.0
tabulate>=0.10.0